+++ NEWS.md
@@ -5,6 +5,7 @@ None
 
 ## New Features
 1. Fix #459 - Add function `ts_growth_rate_vec()`
+2. Fix #463 - Add function `ts_adf_test()`
 
 ## Minor Fixes and Improvements
 1. Fix #456 Fix boilerplate examples to set the `.true` param to `FALSE`